requirements.txt
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2025.4.26
Django==5.2.1
django-phonenumber-field==8.1.0
environs==14.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
marshmallow==4.0.0
phonenumbers==9.0.5
python-dotenv==1.1.0
python-telegram-bot==13.15
pytz==2025.2
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.15
